deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn==1.3.0
plotly
joblib
image
pillow

